Split projects into two projects
