+++ pages/.env.development
@@ -1,0 +1,1 @@
+REACT_APP_WORKERS_HOST=127.0.0.1:8787
