+++ pages/.gitignore
@@ -13,4 +13,11 @@ node_modules
 
 # misc
 .DS_Store
-dist
+.env.local
+.env.development.local
+.env.test.local
+.env.production.local
+
+npm-debug.log*
+yarn-debug.log*
+yarn-error.log*
