+++ pages/public/manifest.json
