+++ workers/package.json
@@ -1,0 +1,18 @@
+{
+  "name": "masquerade",
+  "version": "0.1.0",
+  "private": true,
+  "license": "MIT",
+  "scripts": {
+    "publish": "npx wrangler publish",
+    "build": "esbuild --bundle --sourcemap --format=esm --outdir=dist ./src/index.js",
+    "dev": "miniflare --watch"
+  },
+  "devDependencies": {
+    "@cloudflare/wrangler": "^1.19.5",
+    "esbuild": "^0.14.11"
+  },
+  "dependencies": {
+    "miniflare": "^2.1.0"
+  }
+}
