+++ workers/src/GameRoom.js
@@ -1,0 +1,32 @@
+export class GameRoom {
+  constructor(controller, env) {
+    this.sessions = [];
+  }
+
+  async fetch(request) {
+    let url = new URL(request.url);
+
+    switch (url.pathname) {
+      case "/websocket": {
+        // The request is to `/api/room/<name>/websocket`. A client is trying to establish a new WebSocket session.
+        if (request.headers.get("Upgrade") !== "websocket") {
+          return new Response("expected websocket", { status: 400 });
+        }
+        // eslint-disable-next-line no-undef
+        let pair = new WebSocketPair();
+        await this.handleSession(pair[1]);
+        return new Response(null, { status: 101, webSocket: pair[0] });
+      }
+      default:
+        return new Response("Not found", { status: 404 });
+    }
+  }
+
+  async handleSession(webSocket) {
+    webSocket.accept();
+    webSocket.addEventListener("message", (event) => {
+      console.log(event.data);
+      webSocket.send("PONG");
+    });
+  }
+}
