+++ workers/src/index.js
@@ -8,7 +8,7 @@ function handleWebSocket(request) {
   const [client, server] = Object.values(webSocketPair);
 
   server.accept();
-  new HeartBeatManager({}, server)
+  new HeartBeatManager({}, server);
   server.addEventListener("message", (event) => {
     console.log(event.data);
     server.send("PONG");
@@ -19,6 +19,7 @@ function handleWebSocket(request) {
     webSocket: client,
   });
 }
+export { GameRoom } from "./GameRoom";
 
 // eslint-disable-next-line import/no-anonymous-default-export
 export default {
@@ -26,27 +27,28 @@ export default {
     let url = new URL(request.url);
     let path = url.pathname.slice(1).split("/");
 
-    if (!path[0]) {
-      return env.ASSETS.fetch(request);
-    }
-
     switch (path[0]) {
       case ".functions":
         return handleApiRequest(path.slice(1), request, env);
       default:
-        return env.ASSETS.fetch(request);
+        return new Response("Not Found", { status: 404 });
     }
   },
 };
 
 export class HeartBeatManager {
-  constructor(game, webSocket, checkMilliseconds = 3000, timeoutMilliseconds = 30000) {
+  constructor(
+    game,
+    webSocket,
+    checkMilliseconds = 3000,
+    timeoutMilliseconds = 30000
+  ) {
     this.game = game;
-    this.webSocket = webSocket
+    this.webSocket = webSocket;
     this.checkMilliseconds = checkMilliseconds;
     this.timeoutMilliseconds = timeoutMilliseconds;
     this.heartBeatLastReceivedTime = new Date().getTime();
-    
+
     this.checkIntervalID = setInterval(
       () => this.heartBeatCheck(),
       checkMilliseconds
@@ -58,10 +60,12 @@ export class HeartBeatManager {
         if (message["type"] === "PING") {
           this.heartBeatReceived();
         }
-      } catch(e) { }
+      } catch (e) {}
     });
 
-    this.webSocket.addEventListener("close", (event) => clearInterval(this.checkIntervalID));
+    this.webSocket.addEventListener("close", (event) =>
+      clearInterval(this.checkIntervalID)
+    );
   }
 
   heartBeatReceived() {
@@ -74,9 +78,11 @@ export class HeartBeatManager {
     let now = new Date().getTime();
     if (now - this.heartBeatLastReceivedTime > this.timeoutMilliseconds) {
       // this.game.markPlayerDead();
-      console.log(`${this.webSocket} is dead`)
+      console.log(`${this.webSocket} is dead`);
     }
   }
+}
+
 async function handleApiRequest(path, request, env) {
   switch (path[0]) {
     case "room": {
@@ -106,38 +112,5 @@ async function handleApiRequest(path, request, env) {
     }
     default:
       return new Response("Not found", { status: 404 });
-  }
-}
-
-export class GameRoom {
-  constructor(controller, env) {
-    this.sessions = [];
-  }
-
-  async fetch(request) {
-    let url = new URL(request.url);
-
-    switch (url.pathname) {
-      case "/websocket": {
-        // The request is to `/api/room/<name>/websocket`. A client is trying to establish a new WebSocket session.
-        if (request.headers.get("Upgrade") !== "websocket") {
-          return new Response("expected websocket", { status: 400 });
-        }
-        // eslint-disable-next-line no-undef
-        let pair = new WebSocketPair();
-        await this.handleSession(pair[1]);
-        return new Response(null, { status: 101, webSocket: pair[0] });
-      }
-      default:
-        return new Response("Not found", { status: 404 });
-    }
-  }
-
-  async handleSession(webSocket) {
-    webSocket.accept();
-    webSocket.addEventListener("message", (event) => {
-      console.log(event.data);
-      webSocket.send("PONG");
-    });
   }
 }
