+++ workers/wrangler.toml
@@ -1,0 +1,21 @@
+name = "masquerade-game-room"
+type = "javascript"
+workers_dev = true
+compatibility_date = "2021-11-17"
+
+[build]
+command = "npm run build"
+
+[build.upload]
+dir = "dist"
+format = "modules"
+main = "./index.js"
+
+[durable_objects]
+bindings = [
+  { name = "rooms", class_name = "GameRoom" }
+]
+
+[[migrations]]
+tag = "v1"
+new_classes = ["GameRoom"]
